Day7/Exercise1/requirements.txt
setuptools>=69.0.3
wheel>=0.42.0
numpy>=1.26.0
pandas>=2.2.0
google-cloud-bigquery>=3.13.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.7
python-dotenv>=1.0.0
db-dtypes>=1.1.1
